+++ RUFAS/biophysical/manure/storage/slurry_storage_underfloor.py
@@ -114,22 +114,22 @@ class SlurryStorageUnderfloor(Storage):
             storage_methane_from_degradable_volatile_solids + storage_methane_from_non_degradable_volatile_solids
         )
         bedding_to_manure_non_degradable_volatile_solids_ratio = (
-            self._manure_to_process.bedding_non_degradable_volatile_solids / (
+            self._manure_to_process.bedding_non_degradable_volatile_solids
+            / (
                 self._manure_to_process.non_degradable_volatile_solids
-                + self._manure_to_process.bedding_non_degradable_volatile_solids)
+                + self._manure_to_process.bedding_non_degradable_volatile_solids
+            )
         )
 
         self._manure_to_process.total_solids = max(
             0.0,
-            self._manure_to_process.total_solids
-            - total_storage_methane * ManureConstants.METHANE_TO_METHANE_CARBON_DIOXIDE_RATIO,
+            self._manure_to_process.total_solids - total_storage_methane * ManureConstants.VS_TO_METHANE_LOSS_RATIO,
         )
         self._manure_to_process.degradable_volatile_solids = max(
             0.0,
             (
                 self._manure_to_process.degradable_volatile_solids
-                - storage_methane_from_degradable_volatile_solids
-                * ManureConstants.METHANE_TO_METHANE_CARBON_DIOXIDE_RATIO
+                - storage_methane_from_degradable_volatile_solids * ManureConstants.VS_TO_METHANE_LOSS_RATIO
             ),
         )
         self._manure_to_process.non_degradable_volatile_solids = max(
@@ -137,7 +137,7 @@ class SlurryStorageUnderfloor(Storage):
             self._manure_to_process.non_degradable_volatile_solids
             - (
                 storage_methane_from_non_degradable_volatile_solids
-                * ManureConstants.METHANE_TO_METHANE_CARBON_DIOXIDE_RATIO
+                * ManureConstants.VS_TO_METHANE_LOSS_RATIO
                 * (1 - bedding_to_manure_non_degradable_volatile_solids_ratio)
             ),
         )
@@ -146,7 +146,7 @@ class SlurryStorageUnderfloor(Storage):
             self._manure_to_process.bedding_non_degradable_volatile_solids
             - (
                 storage_methane_from_non_degradable_volatile_solids
-                * ManureConstants.METHANE_TO_METHANE_CARBON_DIOXIDE_RATIO
+                * ManureConstants.VS_TO_METHANE_LOSS_RATIO
                 * bedding_to_manure_non_degradable_volatile_solids_ratio
             ),
         )